Modernize timer classes: unsigned long to int64_t, sentinel to bool
Replace 0xFFFFFFFF magic sentinel with an explicit bool Active flag in
TTimerClass and CDTimerClass. Migrate all timer value types from
unsigned long to int64_t per project integer type conventions.

Add operator=(int64_t) to TTimerClass and CDTimerClass so the common
`timer = value` reset pattern no longer relies on implicit converting
constructors. Clean up Westwood block comments to Google style.

Update PentiumTimerClass to return int64_t and use the modern
Get_CPU_Clock() overload. Propagate int64_t to Benchmark, SpareTicks,
ProcessTimer/ProcessTicks, and all local variables storing timer values.

+++ src/ra/conquer.cc
@@ -4632,7 +4632,7 @@ void Shake_The_Screen(int shakes) {
   int oldyoff = 0;
   int newyoff = 0;
   while (shakes--) {
-    int x = TickCount.Value();
+    int64_t x = TickCount.Value();
     //		CountDownTimer = 1;
     do {
       newyoff = Sim_Random_Pick(0, 2) - 1;
